# 3° Anno/Programmazione per il Web/Progetto/WASAText/doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: una piattaforma di messaggistica che supporta chat individuali e di gruppo, accessibile tramite PC
  version: 1.0.0
...
paths:
...
  /users:
    # Creazione e autenticazione utente
    post:
      tags:
        - login
      summary: Login utente
      description: |-
        Se l'utente non esiste, verrà creato,
        e viene restituito un identificatore.
        Se l'utente esiste, viene restituito l'identificatore dell'utente.
      operationId: doLogin
      requestBody:
        description: dettagli utente
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest' 
      responses:
        '201':
          description: Login utente riuscito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 1.1 - Ricerca utenti mediante username
...
components:
...
  schemas:
...
    User:
      type: object
      description: Rappresenta un utente nella piattaforma WASAText
      required:
        - name
      properties:
        id:
          type: string
          minLength: 6
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: "abcdef012345"
          description: Identificatore univoco dell'utente
        name:
          type: string
          pattern: '^[a-zA-Z0-9._-]{1,64}$'
          example: "Maria"
          minLength: 1
          maxLength: 64
          description: Nome utente univoco
        photo_url:
          type: string
          format: uri
          description: URL della foto profilo dell'utente
          example: "http://example.com/photos/user123.png"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
    LoginUserRequest:
      type: object
      description: Dati necessari per il login o la creazione di un nuovo utente
      required:
        - name
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z0-9._-]{1,64}$'
          example: "Maria"
          minLength: 1
          maxLength: 64
          description: Nome utente univoco


    # Schemi per la chat
    # Inizio creando uno schema che funge da superclasse con proprietà comuni a tutte le chat
...
    Error:
      type: object
      description: Rappresenta un errore generico restituito dalle API
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[A-Z_]{3,64}$'
          example: "NOT_FOUND"
          description: Codice dell'errore
        message:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[\s\S]{3,64}$'
          example: "La risorsa richiesta non è stata trovata."
          description: Descrizione dettagliata dell'errore

  # RESPONSES

  # Risposte per gli errori
  responses:
    BadRequest:
      description: Richiesta non valida (400 Bad Request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Errore interno del server (500 Internal Server Error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # PARAMETERS
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT